Update README.md
Unofficial  BGMI Website is an exciting front-end project that aims to provide a brief information
about the BATTLEGROUNDS MOBILE INDIA  to the users by offering a feature-rich, visually appealing, and user-friendly platform. It brings together the power of modern web development technologies to create a seamless and delightful experience for game enthusiasts.